Refactor import paths for POKEMON_DATA
- Updated import statements across multiple files to source POKEMON_DATA from '../common/dex/index' instead of '../common/pokemon-data'.
- This change standardizes the import location for better organization and maintainability of the codebase.
