+++ src/common/dex/gen1.ts
@@ -1,0 +1,1075 @@
+import { PokemonConfig, PokemonGeneration, PokemonColor } from '../types';
+
+export const GEN_1_DEX: { [key: string]: PokemonConfig } = {
+  bulbasaur: {
+    id: 1,
+    name: 'Bulbasaur',
+    generation: PokemonGeneration.Gen1,
+    cry: 'Bulbasaur!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  ivysaur: {
+    id: 2,
+    name: 'Ivysaur',
+    generation: PokemonGeneration.Gen1,
+    cry: 'Ivysaur!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  venusaur: {
+    id: 3,
+    name: 'Venusaur',
+    generation: PokemonGeneration.Gen1,
+    cry: 'Venusaur!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  venusaur_female: {
+    id: 3,
+    name: 'Venusaur (Female)',
+    generation: PokemonGeneration.Gen1,
+    cry: 'Venusaur!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  charmander: {
+    id: 4,
+    name: 'Charmander',
+    generation: PokemonGeneration.Gen1,
+    cry: 'Charmander!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  charmeleon: {
+    id: 5,
+    name: 'Charmeleon',
+    generation: PokemonGeneration.Gen1,
+    cry: 'Charmeleon!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  charizard: {
+    id: 6,
+    name: 'Charizard',
+    generation: PokemonGeneration.Gen1,
+    cry: 'Charizard!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  squirtle: {
+    id: 7,
+    name: 'Squirtle',
+    generation: PokemonGeneration.Gen1,
+    cry: 'Squritle!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  wartortle: {
+    id: 8,
+    name: 'Wartortle',
+    generation: PokemonGeneration.Gen1,
+    cry: 'Wartortle!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  blastoise: {
+    id: 9,
+    name: 'Blastoise',
+    generation: PokemonGeneration.Gen1,
+    cry: 'Blastoise!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  caterpie: {
+    id: 10,
+    name: 'Caterpie',
+    generation: PokemonGeneration.Gen1,
+    cry: 'Caterpie!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  metapod: {
+    id: 11,
+    name: 'Metapod',
+    generation: PokemonGeneration.Gen1,
+    cry: 'Metapod!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  butterfree: {
+    id: 12,
+    name: 'Butterfree',
+    generation: PokemonGeneration.Gen1,
+    cry: 'Butterfree!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  weedle: {
+    id: 13,
+    name: 'Weedle',
+    generation: PokemonGeneration.Gen1,
+    cry: 'Weedle!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  kakuna: {
+    id: 14,
+    name: 'Kakuna',
+    generation: PokemonGeneration.Gen1,
+    cry: 'Kakuna!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  beedrill: {
+    id: 15,
+    name: 'Beedrill',
+    generation: PokemonGeneration.Gen1,
+    cry: 'Beedrill!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  pidgey: {
+    id: 16,
+    name: 'Pidgey',
+    generation: PokemonGeneration.Gen1,
+    cry: 'Pidgey!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  pidgeotto: {
+    id: 17,
+    name: 'Pidgeotto',
+    generation: PokemonGeneration.Gen1,
+    cry: 'Pidgeotto!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  pidgeot: {
+    id: 18,
+    name: 'Pidgeot',
+    generation: PokemonGeneration.Gen1,
+    cry: 'Pidgeot!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  rattata: {
+    id: 19,
+    name: 'Rattata',
+    generation: PokemonGeneration.Gen1,
+    cry: 'Rattata!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  raticate: {
+    id: 20,
+    name: 'Raticate',
+    generation: PokemonGeneration.Gen1,
+    cry: 'Raticate!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  spearow: {
+    id: 21,
+    name: 'Spearow',
+    generation: PokemonGeneration.Gen1,
+    cry: 'Spearow!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  fearow: {
+    id: 22,
+    name: 'Fearow',
+    generation: PokemonGeneration.Gen1,
+    cry: 'Fearow!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  ekans: {
+    id: 23,
+    name: 'Ekans',
+    generation: PokemonGeneration.Gen1,
+    cry: 'Ekans!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  arbok: {
+    id: 24,
+    name: 'Arbok',
+    generation: PokemonGeneration.Gen1,
+    cry: 'Arbok!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  pikachu: {
+    id: 25,
+    name: 'Pikachu',
+    generation: PokemonGeneration.Gen1,
+    cry: 'Pikachu!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  pikachu_female: {
+    id: 25,
+    name: 'Pikachu (Female)',
+    generation: PokemonGeneration.Gen1,
+    cry: 'Pikachu!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  raichu: {
+    id: 26,
+    name: 'Raichu',
+    generation: PokemonGeneration.Gen1,
+    cry: 'Raichu!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  sandshrew: {
+    id: 27,
+    name: 'Sandshrew',
+    generation: PokemonGeneration.Gen1,
+    cry: 'Sandshrew!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  sandslash: {
+    id: 28,
+    name: 'Sandslash',
+    generation: PokemonGeneration.Gen1,
+    cry: 'Sandslash!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  nidoran_female: {
+    id: 29,
+    name: 'Nidoran♀',
+    generation: PokemonGeneration.Gen1,
+    cry: 'Nidoran!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  nidorina: {
+    id: 30,
+    name: 'Nidorina',
+    generation: PokemonGeneration.Gen1,
+    cry: 'Nidorina!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  nidoqueen: {
+    id: 31,
+    name: 'Nidoqueen',
+    generation: PokemonGeneration.Gen1,
+    cry: 'Nidoqueen!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  nidoran_male: {
+    id: 32,
+    name: 'Nidoran♂',
+    generation: PokemonGeneration.Gen1,
+    cry: 'Nidoran!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  nidorino: {
+    id: 33,
+    name: 'Nidorino',
+    generation: PokemonGeneration.Gen1,
+    cry: 'Nidorino!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  nidoking: {
+    id: 34,
+    name: 'Nidoking',
+    generation: PokemonGeneration.Gen1,
+    cry: 'Nidoking!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  clefairy: {
+    id: 35,
+    name: 'Clefairy',
+    generation: PokemonGeneration.Gen1,
+    cry: 'Clefairy!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  clefable: {
+    id: 36,
+    name: 'Clefable',
+    generation: PokemonGeneration.Gen1,
+    cry: 'Clefable!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  vulpix: {
+    id: 37,
+    name: 'Vulpix',
+    generation: PokemonGeneration.Gen1,
+    cry: 'Vulpix!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  ninetales: {
+    id: 38,
+    name: 'Ninetales',
+    generation: PokemonGeneration.Gen1,
+    cry: 'Ninetales!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  jigglypuff: {
+    id: 39,
+    name: 'Jigglypuff',
+    generation: PokemonGeneration.Gen1,
+    cry: 'Jigglypuff!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  wigglytuff: {
+    id: 40,
+    name: 'Wigglytuff',
+    generation: PokemonGeneration.Gen1,
+    cry: 'Wigglytuff!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  zubat: {
+    id: 41,
+    name: 'Zubat',
+    generation: PokemonGeneration.Gen1,
+    cry: 'Zubat!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  golbat: {
+    id: 42,
+    name: 'Golbat',
+    generation: PokemonGeneration.Gen1,
+    cry: 'Golbat!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  oddish: {
+    id: 43,
+    name: 'Oddish',
+    generation: PokemonGeneration.Gen1,
+    cry: 'Oddish!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  gloom: {
+    id: 44,
+    name: 'Gloom',
+    generation: PokemonGeneration.Gen1,
+    cry: 'Gloom!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  vileplume: {
+    id: 45,
+    name: 'Vileplume',
+    generation: PokemonGeneration.Gen1,
+    cry: 'Vileplume!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  paras: {
+    id: 46,
+    name: 'Paras',
+    generation: PokemonGeneration.Gen1,
+    cry: 'Paras!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  parasect: {
+    id: 47,
+    name: 'Parasect',
+    generation: PokemonGeneration.Gen1,
+    cry: 'Parasect!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  venonat: {
+    id: 48,
+    name: 'Venonat',
+    generation: PokemonGeneration.Gen1,
+    cry: 'Venonat!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  venomoth: {
+    id: 49,
+    name: 'Venomoth',
+    generation: PokemonGeneration.Gen1,
+    cry: 'Venomoth!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  diglett: {
+    id: 50,
+    name: 'Diglett',
+    generation: PokemonGeneration.Gen1,
+    cry: 'Diglett!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  dugtrio: {
+    id: 51,
+    name: 'Dugtrio',
+    generation: PokemonGeneration.Gen1,
+    cry: 'Dugtrio!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  meowth: {
+    id: 52,
+    name: 'Meowth',
+    generation: PokemonGeneration.Gen1,
+    cry: 'Meowth!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  persian: {
+    id: 53,
+    name: 'Persian',
+    generation: PokemonGeneration.Gen1,
+    cry: 'Persian!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  psyduck: {
+    id: 54,
+    name: 'Psyduck',
+    generation: PokemonGeneration.Gen1,
+    cry: 'Psyduck!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  golduck: {
+    id: 55,
+    name: 'Golduck',
+    generation: PokemonGeneration.Gen1,
+    cry: 'Golduck!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  mankey: {
+    id: 56,
+    name: 'Mankey',
+    generation: PokemonGeneration.Gen1,
+    cry: 'Mankey!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  primeape: {
+    id: 57,
+    name: 'Primeape',
+    generation: PokemonGeneration.Gen1,
+    cry: 'Primeape!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  growlithe: {
+    id: 58,
+    name: 'Growlithe',
+    generation: PokemonGeneration.Gen1,
+    cry: 'Growlithe!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  arcanine: {
+    id: 59,
+    name: 'Arcanine',
+    generation: PokemonGeneration.Gen1,
+    cry: 'Arcanine!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  poliwag: {
+    id: 60,
+    name: 'Poliwag',
+    generation: PokemonGeneration.Gen1,
+    cry: 'Poliwag!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  poliwhirl: {
+    id: 61,
+    name: 'Poliwhirl',
+    generation: PokemonGeneration.Gen1,
+    cry: 'Poliwhirl!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  poliwrath: {
+    id: 62,
+    name: 'Poliwrath',
+    generation: PokemonGeneration.Gen1,
+    cry: 'Poliwrath!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  abra: {
+    id: 63,
+    name: 'Abra',
+    generation: PokemonGeneration.Gen1,
+    cry: 'Abra!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  kadabra: {
+    id: 64,
+    name: 'Kadabra',
+    generation: PokemonGeneration.Gen1,
+    cry: 'Kadabra!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  alakazam: {
+    id: 65,
+    name: 'Alakazam',
+    generation: PokemonGeneration.Gen1,
+    cry: 'Alakazam!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  machop: {
+    id: 66,
+    name: 'Machop',
+    generation: PokemonGeneration.Gen1,
+    cry: 'Machop!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  machoke: {
+    id: 67,
+    name: 'Machoke',
+    generation: PokemonGeneration.Gen1,
+    cry: 'Machoke!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  machamp: {
+    id: 68,
+    name: 'Machamp',
+    generation: PokemonGeneration.Gen1,
+    cry: 'Alakazam!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  bellsprout: {
+    id: 69,
+    name: 'Bellsprout',
+    generation: PokemonGeneration.Gen1,
+    cry: 'Bellsprout!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  weepinbell: {
+    id: 70,
+    name: 'Weepinbell',
+    generation: PokemonGeneration.Gen1,
+    cry: 'Weepinbell!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  victreebel: {
+    id: 71,
+    name: 'Victreebel',
+    generation: PokemonGeneration.Gen1,
+    cry: 'Victreebel!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  tentacool: {
+    id: 72,
+    name: 'Tentacool',
+    generation: PokemonGeneration.Gen1,
+    cry: 'Tentacool!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  tentacruel: {
+    id: 73,
+    name: 'Tentacruel',
+    generation: PokemonGeneration.Gen1,
+    cry: 'Tentacruel!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  geodude: {
+    id: 74,
+    name: 'Geodude',
+    generation: PokemonGeneration.Gen1,
+    cry: 'Geodude!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  graveler: {
+    id: 75,
+    name: 'Graveler',
+    generation: PokemonGeneration.Gen1,
+    cry: 'Graveler!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  golem: {
+    id: 76,
+    name: 'Golem',
+    generation: PokemonGeneration.Gen1,
+    cry: 'Golem!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  ponyta: {
+    id: 77,
+    name: 'Ponyta',
+    generation: PokemonGeneration.Gen1,
+    cry: 'Ponyta!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  rapidash: {
+    id: 78,
+    name: 'Rapidash',
+    generation: PokemonGeneration.Gen1,
+    cry: 'Rapidash!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  slowpoke: {
+    id: 79,
+    name: 'Slowpoke',
+    generation: PokemonGeneration.Gen1,
+    cry: 'Slowpoke!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  slowbro: {
+    id: 80,
+    name: 'Slowbro',
+    generation: PokemonGeneration.Gen1,
+    cry: 'Slowbro!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  magnemite: {
+    id: 81,
+    name: 'Magnemite',
+    generation: PokemonGeneration.Gen1,
+    cry: 'Magnemite!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  magneton: {
+    id: 82,
+    name: 'Magneton',
+    generation: PokemonGeneration.Gen1,
+    cry: 'Magneton!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  farfetchd: {
+    id: 83,
+    name: "Farfetch'd",
+    generation: PokemonGeneration.Gen1,
+    cry: "Farfetch'd!",
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  doduo: {
+    id: 84,
+    name: 'Doduo',
+    generation: PokemonGeneration.Gen1,
+    cry: 'Doduo!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  dodrio: {
+    id: 85,
+    name: 'Dodrio',
+    generation: PokemonGeneration.Gen1,
+    cry: 'Dodrio!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  seel: {
+    id: 86,
+    name: 'Seel',
+    generation: PokemonGeneration.Gen1,
+    cry: 'Seel!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  dewgong: {
+    id: 87,
+    name: 'Dewgong',
+    generation: PokemonGeneration.Gen1,
+    cry: 'Dewgong!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  grimer: {
+    id: 88,
+    name: 'Grimer',
+    generation: PokemonGeneration.Gen1,
+    cry: 'Grimer!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  muk: {
+    id: 89,
+    name: 'Muk',
+    generation: PokemonGeneration.Gen1,
+    cry: 'Muk!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  shellder: {
+    id: 90,
+    name: 'Shellder',
+    generation: PokemonGeneration.Gen1,
+    cry: 'Shellder!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  cloyster: {
+    id: 91,
+    name: 'Cloyster',
+    generation: PokemonGeneration.Gen1,
+    cry: 'Cloyster!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  gastly: {
+    id: 92,
+    name: 'Gastly',
+    generation: PokemonGeneration.Gen1,
+    cry: 'Gastly!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  haunter: {
+    id: 93,
+    name: 'Haunter',
+    generation: PokemonGeneration.Gen1,
+    cry: 'Haunter!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  gengar: {
+    id: 94,
+    name: 'Gengar',
+    generation: PokemonGeneration.Gen1,
+    cry: 'Gengar!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  onix: {
+    id: 95,
+    name: 'Onix',
+    generation: PokemonGeneration.Gen1,
+    cry: 'Onix!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  drowzee: {
+    id: 96,
+    name: 'Drowzee',
+    generation: PokemonGeneration.Gen1,
+    cry: 'Drowzee!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  hypno: {
+    id: 97,
+    name: 'Hypno',
+    generation: PokemonGeneration.Gen1,
+    cry: 'Hypno!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  krabby: {
+    id: 98,
+    name: 'Krabby',
+    generation: PokemonGeneration.Gen1,
+    cry: 'Krabby!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  kingler: {
+    id: 99,
+    name: 'Kingler',
+    generation: PokemonGeneration.Gen1,
+    cry: 'Kingler!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  voltorb: {
+    id: 100,
+    name: 'Voltorb',
+    generation: PokemonGeneration.Gen1,
+    cry: 'Voltorb!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  electrode: {
+    id: 101,
+    name: 'Electrode',
+    generation: PokemonGeneration.Gen1,
+    cry: 'Electrode!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  exeggcute: {
+    id: 102,
+    name: 'Exeggcute',
+    generation: PokemonGeneration.Gen1,
+    cry: 'Exeggcute!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  exeggutor: {
+    id: 103,
+    name: 'Exeggutor',
+    generation: PokemonGeneration.Gen1,
+    cry: 'Exeggutor!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  cubone: {
+    id: 104,
+    name: 'Cubone',
+    generation: PokemonGeneration.Gen1,
+    cry: 'Cubone!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  marowak: {
+    id: 105,
+    name: 'Marowak',
+    generation: PokemonGeneration.Gen1,
+    cry: 'Marowak!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  hitmonlee: {
+    id: 106,
+    name: 'Hitmonlee',
+    generation: PokemonGeneration.Gen1,
+    cry: 'Hitmonlee!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  hitmonchan: {
+    id: 107,
+    name: 'Hitmonchan',
+    generation: PokemonGeneration.Gen1,
+    cry: 'Hitmonchan!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  lickitung: {
+    id: 108,
+    name: 'Lickitung',
+    generation: PokemonGeneration.Gen1,
+    cry: 'Lickitung!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  koffing: {
+    id: 109,
+    name: 'Koffing',
+    generation: PokemonGeneration.Gen1,
+    cry: 'Koffing!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  weezing: {
+    id: 110,
+    name: 'Weezing',
+    generation: PokemonGeneration.Gen1,
+    cry: 'Weezing!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  rhyhorn: {
+    id: 111,
+    name: 'Rhyhorn',
+    generation: PokemonGeneration.Gen1,
+    cry: 'Rhyhorn!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  rhydon: {
+    id: 112,
+    name: 'Rhydon',
+    generation: PokemonGeneration.Gen1,
+    cry: 'Rhydon!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  chansey: {
+    id: 113,
+    name: 'Chansey',
+    generation: PokemonGeneration.Gen1,
+    cry: 'Chansey!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  tangela: {
+    id: 114,
+    name: 'Tangela',
+    generation: PokemonGeneration.Gen1,
+    cry: 'Tangela!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  kangaskhan: {
+    id: 115,
+    name: 'Kangaskhan',
+    generation: PokemonGeneration.Gen1,
+    cry: 'Kangaskhan!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  horsea: {
+    id: 116,
+    name: 'Horsea',
+    generation: PokemonGeneration.Gen1,
+    cry: 'Horsea!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  seadra: {
+    id: 117,
+    name: 'Seadra',
+    generation: PokemonGeneration.Gen1,
+    cry: 'Seadra!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  goldeen: {
+    id: 118,
+    name: 'Goldeen',
+    generation: PokemonGeneration.Gen1,
+    cry: 'Goldeen!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  seaking: {
+    id: 119,
+    name: 'Seaking',
+    generation: PokemonGeneration.Gen1,
+    cry: 'Seaking!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  staryu: {
+    id: 120,
+    name: 'Staryu',
+    generation: PokemonGeneration.Gen1,
+    cry: 'Staryu!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  starmie: {
+    id: 121,
+    name: 'Starmie',
+    generation: PokemonGeneration.Gen1,
+    cry: 'Starmie!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  mrmime: {
+    id: 122,
+    name: 'Mr. Mime',
+    generation: PokemonGeneration.Gen1,
+    cry: 'Mr. Mime!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  scyther: {
+    id: 123,
+    name: 'Scyther',
+    generation: PokemonGeneration.Gen1,
+    cry: 'Scyther!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  jynx: {
+    id: 124,
+    name: 'Jynx',
+    generation: PokemonGeneration.Gen1,
+    cry: 'Jynx!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  electabuzz: {
+    id: 125,
+    name: 'Electabuzz',
+    generation: PokemonGeneration.Gen1,
+    cry: 'Electabuzz!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  magmar: {
+    id: 126,
+    name: 'Magmar',
+    generation: PokemonGeneration.Gen1,
+    cry: 'Magmar!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  pinsir: {
+    id: 127,
+    name: 'Pinsir',
+    generation: PokemonGeneration.Gen1,
+    cry: 'Pinsir!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  tauros: {
+    id: 128,
+    name: 'Tauros',
+    generation: PokemonGeneration.Gen1,
+    cry: 'Tauros!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  magikarp: {
+    id: 129,
+    name: 'Magikarp',
+    generation: PokemonGeneration.Gen1,
+    cry: 'Magikarp!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  gyarados: {
+    id: 130,
+    name: 'Gyarados',
+    generation: PokemonGeneration.Gen1,
+    cry: 'Gyarados!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  lapras: {
+    id: 131,
+    name: 'Lapras',
+    generation: PokemonGeneration.Gen1,
+    cry: 'Lapras!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  ditto: {
+    id: 132,
+    name: 'Ditto',
+    generation: PokemonGeneration.Gen1,
+    cry: 'Ditto!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  eevee: {
+    id: 133,
+    name: 'Eevee',
+    generation: PokemonGeneration.Gen1,
+    cry: 'Eevee!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  vaporeon: {
+    id: 134,
+    name: 'Vaporeon',
+    generation: PokemonGeneration.Gen1,
+    cry: 'Vaporeon!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  jolteon: {
+    id: 135,
+    name: 'Jolteon',
+    generation: PokemonGeneration.Gen1,
+    cry: 'Jolteon!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  flareon: {
+    id: 136,
+    name: 'Flareon',
+    generation: PokemonGeneration.Gen1,
+    cry: 'Flareon!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  porygon: {
+    id: 137,
+    name: 'Porygon',
+    generation: PokemonGeneration.Gen1,
+    cry: 'Porygon!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  omanyte: {
+    id: 138,
+    name: 'Omanyte',
+    generation: PokemonGeneration.Gen1,
+    cry: 'Omanyte!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  omastar: {
+    id: 139,
+    name: 'Omastar',
+    generation: PokemonGeneration.Gen1,
+    cry: 'Omastar!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  kabuto: {
+    id: 140,
+    name: 'Kabuto',
+    generation: PokemonGeneration.Gen1,
+    cry: 'Kabuto!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  kabutops: {
+    id: 141,
+    name: 'Kabutops',
+    generation: PokemonGeneration.Gen1,
+    cry: 'Kabutops!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  aerodactyl: {
+    id: 142,
+    name: 'Aerodactyl',
+    generation: PokemonGeneration.Gen1,
+    cry: 'Aerodactyl!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  snorlax: {
+    id: 143,
+    name: 'Snorlax',
+    generation: PokemonGeneration.Gen1,
+    cry: 'Snorlax!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  articuno: {
+    id: 144,
+    name: 'Articuno',
+    generation: PokemonGeneration.Gen1,
+    cry: 'Articuno!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  zapdos: {
+    id: 145,
+    name: 'Zapdos',
+    generation: PokemonGeneration.Gen1,
+    cry: 'Zapdos!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  moltres: {
+    id: 146,
+    name: 'Moltres',
+    generation: PokemonGeneration.Gen1,
+    cry: 'Moltres!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  dratini: {
+    id: 147,
+    name: 'Dratini',
+    generation: PokemonGeneration.Gen1,
+    cry: 'Dratini!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  dragonair: {
+    id: 148,
+    name: 'Dragonair',
+    generation: PokemonGeneration.Gen1,
+    cry: 'Dragonair!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  dragonite: {
+    id: 149,
+    name: 'Dragonite',
+    generation: PokemonGeneration.Gen1,
+    cry: 'Dragonite!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  mewtwo: {
+    id: 150,
+    name: 'Mewtwo',
+    generation: PokemonGeneration.Gen1,
+    cry: 'Mewtwo!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  mew: {
+    id: 151,
+    name: 'Mew',
+    generation: PokemonGeneration.Gen1,
+    cry: 'Mew!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+};
