+++ src/common/dex/gen2.ts
@@ -1,0 +1,932 @@
+import { PokemonColor, PokemonConfig, PokemonGeneration } from '../types';
+
+export const GEN_2_DEX: { [key: string]: PokemonConfig } = {
+  chikorita: {
+    id: 152,
+    name: 'Chikorita',
+    generation: PokemonGeneration.Gen2,
+    cry: 'Chikorita!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  bayleef: {
+    id: 153,
+    name: 'Bayleef',
+    generation: PokemonGeneration.Gen2,
+    cry: 'Bayleef!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  meganium: {
+    id: 154,
+    name: 'Meganium',
+    generation: PokemonGeneration.Gen2,
+    cry: 'Meganium!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  meganium_female: {
+    id: 154,
+    name: 'Meganium (Female)',
+    generation: PokemonGeneration.Gen2,
+    cry: 'Meganium!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  cyndaquil: {
+    id: 155,
+    name: 'Cyndaquil',
+    generation: PokemonGeneration.Gen2,
+    cry: 'Cyndaquil!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  quilava: {
+    id: 156,
+    name: 'Quilava',
+    generation: PokemonGeneration.Gen2,
+    cry: 'Quilava!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  typhlosion: {
+    id: 157,
+    name: 'Typhlosion',
+    generation: PokemonGeneration.Gen2,
+    cry: 'Typhlosion!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  totodile: {
+    id: 158,
+    name: 'Totodile',
+    generation: PokemonGeneration.Gen2,
+    cry: 'Totodile!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  croconaw: {
+    id: 159,
+    name: 'Croconaw',
+    generation: PokemonGeneration.Gen2,
+    cry: 'Croconaw!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  feraligatr: {
+    id: 160,
+    name: 'Feraligatr',
+    generation: PokemonGeneration.Gen2,
+    cry: 'Feraligatr!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  sentret: {
+    id: 161,
+    name: 'Sentret',
+    generation: PokemonGeneration.Gen2,
+    cry: 'Sentret!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  furret: {
+    id: 162,
+    name: 'Furret',
+    generation: PokemonGeneration.Gen2,
+    cry: 'Furret!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  hoothoot: {
+    id: 163,
+    name: 'Hoothoot',
+    generation: PokemonGeneration.Gen2,
+    cry: 'Hoothoot!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  noctowl: {
+    id: 164,
+    name: 'Noctowl',
+    generation: PokemonGeneration.Gen2,
+    cry: 'Noctowl!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  ledyba: {
+    id: 165,
+    name: 'Ledyba',
+    generation: PokemonGeneration.Gen2,
+    cry: 'Ledyba!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  ledian: {
+    id: 166,
+    name: 'Ledian',
+    generation: PokemonGeneration.Gen2,
+    cry: 'Ledian!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  spinarak: {
+    id: 167,
+    name: 'Spinarak',
+    generation: PokemonGeneration.Gen2,
+    cry: 'Spinarak!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  ariados: {
+    id: 168,
+    name: 'Ariados',
+    generation: PokemonGeneration.Gen2,
+    cry: 'Ariados!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  crobat: {
+    id: 169,
+    name: 'Crobat',
+    generation: PokemonGeneration.Gen2,
+    cry: 'Crobat!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  chinchou: {
+    id: 170,
+    name: 'Chinchou',
+    generation: PokemonGeneration.Gen2,
+    cry: 'Chinchou!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  lanturn: {
+    id: 171,
+    name: 'Lanturn',
+    generation: PokemonGeneration.Gen2,
+    cry: 'Lanturn!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  pichu: {
+    id: 172,
+    name: 'Pichu',
+    generation: PokemonGeneration.Gen2,
+    cry: 'Pichu!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  pichu_spiky_eared: {
+    id: 172,
+    name: 'Pichu (Spiky-eared)',
+    generation: PokemonGeneration.Gen2,
+    cry: 'Pichu!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  cleffa: {
+    id: 173,
+    name: 'Cleffa',
+    generation: PokemonGeneration.Gen2,
+    cry: 'Cleffa!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  igglybuff: {
+    id: 174,
+    name: 'Igglybuff',
+    generation: PokemonGeneration.Gen2,
+    cry: 'Igglybuff!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  togepi: {
+    id: 175,
+    name: 'Togepi',
+    generation: PokemonGeneration.Gen2,
+    cry: 'Togepi!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  togetic: {
+    id: 176,
+    name: 'Togetic',
+    generation: PokemonGeneration.Gen2,
+    cry: 'Togetic!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  natu: {
+    id: 177,
+    name: 'Natu',
+    generation: PokemonGeneration.Gen2,
+    cry: 'Natu!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  xatu: {
+    id: 178,
+    name: 'Xatu',
+    generation: PokemonGeneration.Gen2,
+    cry: 'Xatu!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  mareep: {
+    id: 179,
+    name: 'Mareep',
+    generation: PokemonGeneration.Gen2,
+    cry: 'Mareep!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  flaaffy: {
+    id: 180,
+    name: 'Flaaffy',
+    generation: PokemonGeneration.Gen2,
+    cry: 'Flaaffy!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  ampharos: {
+    id: 181,
+    name: 'Ampharos',
+    generation: PokemonGeneration.Gen2,
+    cry: 'Ampharos!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  bellossom: {
+    id: 182,
+    name: 'Bellossom',
+    generation: PokemonGeneration.Gen2,
+    cry: 'Bellossom!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  marill: {
+    id: 183,
+    name: 'Marill',
+    generation: PokemonGeneration.Gen2,
+    cry: 'Marill!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  azumarill: {
+    id: 184,
+    name: 'Azumarill',
+    generation: PokemonGeneration.Gen2,
+    cry: 'Azumarill!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  sudowoodo: {
+    id: 185,
+    name: 'Sudowoodo',
+    generation: PokemonGeneration.Gen2,
+    cry: 'Sudowoodo!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  politoed: {
+    id: 186,
+    name: 'Politoed',
+    generation: PokemonGeneration.Gen2,
+    cry: 'Politoed!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  hoppip: {
+    id: 187,
+    name: 'Hoppip',
+    generation: PokemonGeneration.Gen2,
+    cry: 'Hoppip!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  skiploom: {
+    id: 188,
+    name: 'Skiploom',
+    generation: PokemonGeneration.Gen2,
+    cry: 'Skiploom!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  jumpluff: {
+    id: 189,
+    name: 'Jumpluff',
+    generation: PokemonGeneration.Gen2,
+    cry: 'Jumpluff!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  aipom: {
+    id: 190,
+    name: 'Aipom',
+    generation: PokemonGeneration.Gen2,
+    cry: 'Aipom!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  sunkern: {
+    id: 191,
+    name: 'Sunkern',
+    generation: PokemonGeneration.Gen2,
+    cry: 'Sunkern!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  sunflora: {
+    id: 192,
+    name: 'Sunflora',
+    generation: PokemonGeneration.Gen2,
+    cry: 'Sunflora!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  yanma: {
+    id: 193,
+    name: 'Yanma',
+    generation: PokemonGeneration.Gen2,
+    cry: 'Yanma!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  wooper: {
+    id: 194,
+    name: 'Wooper',
+    generation: PokemonGeneration.Gen2,
+    cry: 'Wooper!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  quagsire: {
+    id: 195,
+    name: 'Quagsire',
+    generation: PokemonGeneration.Gen2,
+    cry: 'Quagsire!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  espeon: {
+    id: 196,
+    name: 'Espeon',
+    generation: PokemonGeneration.Gen2,
+    cry: 'Espeon!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  umbreon: {
+    id: 197,
+    name: 'Umbreon',
+    generation: PokemonGeneration.Gen2,
+    cry: 'Umbreon!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  murkrow: {
+    id: 198,
+    name: 'Murkrow',
+    generation: PokemonGeneration.Gen2,
+    cry: 'Murkrow!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  slowking: {
+    id: 199,
+    name: 'Slowking',
+    generation: PokemonGeneration.Gen2,
+    cry: 'Slowking!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  misdreavus: {
+    id: 200,
+    name: 'Misdreavus',
+    generation: PokemonGeneration.Gen2,
+    cry: 'Misdreavus!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  unown_a: {
+    id: 201,
+    name: 'Unown A',
+    generation: PokemonGeneration.Gen2,
+    cry: 'Unown!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  unown_b: {
+    id: 201,
+    name: 'Unown B',
+    generation: PokemonGeneration.Gen2,
+    cry: 'Unown!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  unown_c: {
+    id: 201,
+    name: 'Unown C',
+    generation: PokemonGeneration.Gen2,
+    cry: 'Unown!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  unown_d: {
+    id: 201,
+    name: 'Unown D',
+    generation: PokemonGeneration.Gen2,
+    cry: 'Unown!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  unown_e: {
+    id: 201,
+    name: 'Unown E',
+    generation: PokemonGeneration.Gen2,
+    cry: 'Unown!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  unown_f: {
+    id: 201,
+    name: 'Unown F',
+    generation: PokemonGeneration.Gen2,
+    cry: 'Unown!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  unown_g: {
+    id: 201,
+    name: 'Unown G',
+    generation: PokemonGeneration.Gen2,
+    cry: 'Unown!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  unown_h: {
+    id: 201,
+    name: 'Unown H',
+    generation: PokemonGeneration.Gen2,
+    cry: 'Unown!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  unown_i: {
+    id: 201,
+    name: 'Unown I',
+    generation: PokemonGeneration.Gen2,
+    cry: 'Unown!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  unown_j: {
+    id: 201,
+    name: 'Unown J',
+    generation: PokemonGeneration.Gen2,
+    cry: 'Unown!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  unown_k: {
+    id: 201,
+    name: 'Unown K',
+    generation: PokemonGeneration.Gen2,
+    cry: 'Unown!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  unown_l: {
+    id: 201,
+    name: 'Unown L',
+    generation: PokemonGeneration.Gen2,
+    cry: 'Unown!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  unown_m: {
+    id: 201,
+    name: 'Unown M',
+    generation: PokemonGeneration.Gen2,
+    cry: 'Unown!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  unown_n: {
+    id: 201,
+    name: 'Unown N',
+    generation: PokemonGeneration.Gen2,
+    cry: 'Unown!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  unown_o: {
+    id: 201,
+    name: 'Unown O',
+    generation: PokemonGeneration.Gen2,
+    cry: 'Unown!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  unown_p: {
+    id: 201,
+    name: 'Unown P',
+    generation: PokemonGeneration.Gen2,
+    cry: 'Unown!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  unown_q: {
+    id: 201,
+    name: 'Unown Q',
+    generation: PokemonGeneration.Gen2,
+    cry: 'Unown!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  unown_r: {
+    id: 201,
+    name: 'Unown R',
+    generation: PokemonGeneration.Gen2,
+    cry: 'Unown!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  unown_s: {
+    id: 201,
+    name: 'Unown S',
+    generation: PokemonGeneration.Gen2,
+    cry: 'Unown!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  unown_t: {
+    id: 201,
+    name: 'Unown T',
+    generation: PokemonGeneration.Gen2,
+    cry: 'Unown!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  unown_u: {
+    id: 201,
+    name: 'Unown U',
+    generation: PokemonGeneration.Gen2,
+    cry: 'Unown!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  unown_v: {
+    id: 201,
+    name: 'Unown V',
+    generation: PokemonGeneration.Gen2,
+    cry: 'Unown!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  unown_w: {
+    id: 201,
+    name: 'Unown W',
+    generation: PokemonGeneration.Gen2,
+    cry: 'Unown!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  unown_x: {
+    id: 201,
+    name: 'Unown X',
+    generation: PokemonGeneration.Gen2,
+    cry: 'Unown!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  unown_y: {
+    id: 201,
+    name: 'Unown Y',
+    generation: PokemonGeneration.Gen2,
+    cry: 'Unown!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  unown_z: {
+    id: 201,
+    name: 'Unown Z',
+    generation: PokemonGeneration.Gen2,
+    cry: 'Unown!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  unown_exclamation: {
+    id: 201,
+    name: 'Unown !',
+    generation: PokemonGeneration.Gen2,
+    cry: 'Unown!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  unown_question: {
+    id: 201,
+    name: 'Unown ?',
+    generation: PokemonGeneration.Gen2,
+    cry: 'Unown!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  wobbuffet_male: {
+    id: 202,
+    name: 'Wobbuffet (Male)',
+    generation: PokemonGeneration.Gen2,
+    cry: 'Wobbuffet!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  wobbuffet_female: {
+    id: 202,
+    name: 'Wobbuffet (Female)',
+    generation: PokemonGeneration.Gen2,
+    cry: 'Wobbuffet!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  girafarig: {
+    id: 203,
+    name: 'Girafarig',
+    generation: PokemonGeneration.Gen2,
+    cry: 'Girafarig!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  pineco: {
+    id: 204,
+    name: 'Pineco',
+    generation: PokemonGeneration.Gen2,
+    cry: 'Pineco!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  forretress: {
+    id: 205,
+    name: 'Forretress',
+    generation: PokemonGeneration.Gen2,
+    cry: 'Forretress!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  dunsparce: {
+    id: 206,
+    name: 'Dunsparce',
+    generation: PokemonGeneration.Gen2,
+    cry: 'Dunsparce!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  gligar: {
+    id: 207,
+    name: 'Gligar',
+    generation: PokemonGeneration.Gen2,
+    cry: 'Gligar!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  steelix: {
+    id: 208,
+    name: 'Steelix',
+    generation: PokemonGeneration.Gen2,
+    cry: 'Steelix!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+    originalSpriteSize: 64,
+  },
+  steelix_female: {
+    id: 208,
+    name: 'Steelix (Female)',
+    generation: PokemonGeneration.Gen2,
+    cry: 'Steelix!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+    originalSpriteSize: 64,
+  },
+  snubbull: {
+    id: 209,
+    name: 'Snubbull',
+    generation: PokemonGeneration.Gen2,
+    cry: 'Snubbull!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  granbull: {
+    id: 210,
+    name: 'Granbull',
+    generation: PokemonGeneration.Gen2,
+    cry: 'Granbull!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  qwilfish: {
+    id: 211,
+    name: 'Qwilfish',
+    generation: PokemonGeneration.Gen2,
+    cry: 'Qwilfish!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  scizor: {
+    id: 212,
+    name: 'Scizor',
+    generation: PokemonGeneration.Gen2,
+    cry: 'Scizor!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  shuckle: {
+    id: 213,
+    name: 'Shuckle',
+    generation: PokemonGeneration.Gen2,
+    cry: 'Shuckle!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  heracross_male: {
+    id: 214,
+    name: 'Heracross (Male)',
+    generation: PokemonGeneration.Gen2,
+    cry: 'Heracross!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  heracross_female: {
+    id: 214,
+    name: 'Heracross (Female)',
+    generation: PokemonGeneration.Gen2,
+    cry: 'Heracross!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  sneasel: {
+    id: 215,
+    name: 'Sneasel',
+    generation: PokemonGeneration.Gen2,
+    cry: 'Sneasel!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  teddiursa: {
+    id: 216,
+    name: 'Teddiursa',
+    generation: PokemonGeneration.Gen2,
+    cry: 'Teddiursa!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  ursaring: {
+    id: 217,
+    name: 'Ursaring',
+    generation: PokemonGeneration.Gen2,
+    cry: 'Ursaring!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  slugma: {
+    id: 218,
+    name: 'Slugma',
+    generation: PokemonGeneration.Gen2,
+    cry: 'Slugma!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  magcargo: {
+    id: 219,
+    name: 'Magcargo',
+    generation: PokemonGeneration.Gen2,
+    cry: 'Magcargo!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  swinub: {
+    id: 220,
+    name: 'Swinub',
+    generation: PokemonGeneration.Gen2,
+    cry: 'Swinub!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  piloswine: {
+    id: 221,
+    name: 'Piloswine',
+    generation: PokemonGeneration.Gen2,
+    cry: 'Piloswine!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  corsola: {
+    id: 222,
+    name: 'Corsola',
+    generation: PokemonGeneration.Gen2,
+    cry: 'Corsola!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  remoraid: {
+    id: 223,
+    name: 'Remoraid',
+    generation: PokemonGeneration.Gen2,
+    cry: 'Remoraid!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  octillery: {
+    id: 224,
+    name: 'Octillery',
+    generation: PokemonGeneration.Gen2,
+    cry: 'Octillery!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  delibird: {
+    id: 225,
+    name: 'Delibird',
+    generation: PokemonGeneration.Gen2,
+    cry: 'Delibird!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  mantine: {
+    id: 226,
+    name: 'Mantine',
+    generation: PokemonGeneration.Gen2,
+    cry: 'Mantine!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  skarmory: {
+    id: 227,
+    name: 'Skarmory',
+    generation: PokemonGeneration.Gen2,
+    cry: 'Skarmory!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  houndour: {
+    id: 228,
+    name: 'Houndour',
+    generation: PokemonGeneration.Gen2,
+    cry: 'Houndour!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  houndoom: {
+    id: 229,
+    name: 'Houndoom',
+    generation: PokemonGeneration.Gen2,
+    cry: 'Houndoom!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  kingdra: {
+    id: 230,
+    name: 'Kingdra',
+    generation: PokemonGeneration.Gen2,
+    cry: 'Kingdra!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  phanpy: {
+    id: 231,
+    name: 'Phanpy',
+    generation: PokemonGeneration.Gen2,
+    cry: 'Phanpy!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  donphan: {
+    id: 232,
+    name: 'Donphan',
+    generation: PokemonGeneration.Gen2,
+    cry: 'Donphan!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  porygon2: {
+    id: 233,
+    name: 'Porygon2',
+    generation: PokemonGeneration.Gen2,
+    cry: 'Porygon2!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  stantler: {
+    id: 234,
+    name: 'Stantler',
+    generation: PokemonGeneration.Gen2,
+    cry: 'Stantler!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  smeargle: {
+    id: 235,
+    name: 'Smeargle',
+    generation: PokemonGeneration.Gen2,
+    cry: 'Smeargle!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  tyrogue: {
+    id: 236,
+    name: 'Tyrogue',
+    generation: PokemonGeneration.Gen2,
+    cry: 'Tyrogue!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  hitmontop: {
+    id: 237,
+    name: 'Hitmontop',
+    generation: PokemonGeneration.Gen2,
+    cry: 'Hitmontop!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  smoochum: {
+    id: 238,
+    name: 'Smoochum',
+    generation: PokemonGeneration.Gen2,
+    cry: 'Smoochum!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  elekid: {
+    id: 239,
+    name: 'Elekid',
+    generation: PokemonGeneration.Gen2,
+    cry: 'Elekid!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  magby: {
+    id: 240,
+    name: 'Magby',
+    generation: PokemonGeneration.Gen2,
+    cry: 'Magby!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  miltank: {
+    id: 241,
+    name: 'Miltank',
+    generation: PokemonGeneration.Gen2,
+    cry: 'Miltank!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  blissey: {
+    id: 242,
+    name: 'Blissey',
+    generation: PokemonGeneration.Gen2,
+    cry: 'Blissey!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  raikou: {
+    id: 243,
+    name: 'Raikou',
+    generation: PokemonGeneration.Gen2,
+    cry: 'Raikou!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  entei: {
+    id: 244,
+    name: 'Entei',
+    generation: PokemonGeneration.Gen2,
+    cry: 'Entei!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  suicune: {
+    id: 245,
+    name: 'Suicune',
+    generation: PokemonGeneration.Gen2,
+    cry: 'Suicune!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  larvitar: {
+    id: 246,
+    name: 'Larvitar',
+    generation: PokemonGeneration.Gen2,
+    cry: 'Larvitar!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  pupitar: {
+    id: 247,
+    name: 'Pupitar',
+    generation: PokemonGeneration.Gen2,
+    cry: 'Pupitar!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  tyranitar: {
+    id: 248,
+    name: 'Tyranitar',
+    generation: PokemonGeneration.Gen2,
+    cry: 'Tyranitar!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  lugia: {
+    id: 249,
+    name: 'Lugia',
+    generation: PokemonGeneration.Gen2,
+    cry: 'Lugia!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+    originalSpriteSize: 64,
+  },
+  hooh: {
+    id: 250,
+    name: 'Ho-Oh',
+    generation: PokemonGeneration.Gen2,
+    cry: 'Ho-Oh!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+    originalSpriteSize: 64,
+  },
+  celebi: {
+    id: 251,
+    name: 'Celebi',
+    generation: PokemonGeneration.Gen2,
+    cry: 'Celebi!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+};
