+++ src/common/dex/gen3.ts
@@ -1,0 +1,982 @@
+import {
+  PokemonConfig,
+  PokemonGeneration,
+  PokemonColor,
+  PokemonExtraSprite,
+} from '../types';
+
+export const GEN_3_DEX: { [key: string]: PokemonConfig } = {
+  treecko: {
+    id: 252,
+    name: 'Treecko',
+    generation: PokemonGeneration.Gen3,
+    cry: 'Treecko!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  grovyle: {
+    id: 253,
+    name: 'Grovyle',
+    generation: PokemonGeneration.Gen3,
+    cry: 'Grovyle!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  sceptile: {
+    id: 254,
+    name: 'Sceptile',
+    generation: PokemonGeneration.Gen3,
+    cry: 'Sceptile!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  torchic: {
+    id: 255,
+    name: 'Torchic',
+    generation: PokemonGeneration.Gen3,
+    cry: 'Torchic!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  combusken: {
+    id: 256,
+    name: 'Combusken',
+    generation: PokemonGeneration.Gen3,
+    cry: 'Combusken!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  blaziken: {
+    id: 257,
+    name: 'Blaziken',
+    generation: PokemonGeneration.Gen3,
+    cry: 'Blaziken!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  mudkip: {
+    id: 258,
+    name: 'Mudkip',
+    generation: PokemonGeneration.Gen3,
+    cry: 'Mudkip!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  marshtomp: {
+    id: 259,
+    name: 'Marshtomp',
+    generation: PokemonGeneration.Gen3,
+    cry: 'Marshtomp!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  swampert: {
+    id: 260,
+    name: 'Swampert',
+    generation: PokemonGeneration.Gen3,
+    cry: 'Swampert!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  poochyena: {
+    id: 261,
+    name: 'Poochyena',
+    generation: PokemonGeneration.Gen3,
+    cry: 'Poochyena!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  mightyena: {
+    id: 262,
+    name: 'Mightyena',
+    generation: PokemonGeneration.Gen3,
+    cry: 'Mightyena!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  zigzagoon: {
+    id: 263,
+    name: 'Zigzagoon',
+    generation: PokemonGeneration.Gen3,
+    cry: 'Zigzagoon!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  linoone: {
+    id: 264,
+    name: 'Linoone',
+    generation: PokemonGeneration.Gen3,
+    cry: 'Linoone!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  wurmple: {
+    id: 265,
+    name: 'Wurmple',
+    generation: PokemonGeneration.Gen3,
+    cry: 'Wurmple!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  silcoon: {
+    id: 266,
+    name: 'Silcoon',
+    generation: PokemonGeneration.Gen3,
+    cry: 'Silcoon!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  beautifly: {
+    id: 267,
+    name: 'Beautifly',
+    generation: PokemonGeneration.Gen3,
+    cry: 'Beautifly!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  cascoon: {
+    id: 268,
+    name: 'Cascoon',
+    generation: PokemonGeneration.Gen3,
+    cry: 'Cascoon!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  dustox: {
+    id: 269,
+    name: 'Dustox',
+    generation: PokemonGeneration.Gen3,
+    cry: 'Dustox!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  lotad: {
+    id: 270,
+    name: 'Lotad',
+    generation: PokemonGeneration.Gen3,
+    cry: 'Lotad!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  lombre: {
+    id: 271,
+    name: 'Lombre',
+    generation: PokemonGeneration.Gen3,
+    cry: 'Lombre!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  ludicolo: {
+    id: 272,
+    name: 'Ludicolo',
+    generation: PokemonGeneration.Gen3,
+    cry: 'Ludicolo!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  seedot: {
+    id: 273,
+    name: 'Seedot',
+    generation: PokemonGeneration.Gen3,
+    cry: 'Seedot!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  nuzleaf: {
+    id: 274,
+    name: 'Nuzleaf',
+    generation: PokemonGeneration.Gen3,
+    cry: 'Nuzleaf!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  shiftry: {
+    id: 275,
+    name: 'Shiftry',
+    generation: PokemonGeneration.Gen3,
+    cry: 'Shiftry!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  taillow: {
+    id: 276,
+    name: 'Taillow',
+    generation: PokemonGeneration.Gen3,
+    cry: 'Taillow!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  swellow: {
+    id: 277,
+    name: 'Swellow',
+    generation: PokemonGeneration.Gen3,
+    cry: 'Swellow!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  wingull: {
+    id: 278,
+    name: 'Wingull',
+    generation: PokemonGeneration.Gen3,
+    cry: 'Wingull!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  pelipper: {
+    id: 279,
+    name: 'Pelipper',
+    generation: PokemonGeneration.Gen3,
+    cry: 'Pelipper!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  ralts: {
+    id: 280,
+    name: 'Ralts',
+    generation: PokemonGeneration.Gen3,
+    cry: 'Ralts!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  kirlia: {
+    id: 281,
+    name: 'Kirlia',
+    generation: PokemonGeneration.Gen3,
+    cry: 'Kirlia!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  gardevoir: {
+    id: 282,
+    name: 'Gardevoir',
+    generation: PokemonGeneration.Gen3,
+    cry: 'Gardevoir!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  surskit: {
+    id: 283,
+    name: 'Surskit',
+    generation: PokemonGeneration.Gen3,
+    cry: 'Surskit!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  masquerain: {
+    id: 284,
+    name: 'Masquerain',
+    generation: PokemonGeneration.Gen3,
+    cry: 'Masquerain!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  shroomish: {
+    id: 285,
+    name: 'Shroomish',
+    generation: PokemonGeneration.Gen3,
+    cry: 'Shroomish!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  breloom: {
+    id: 286,
+    name: 'Breloom',
+    generation: PokemonGeneration.Gen3,
+    cry: 'Breloom!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  slakoth: {
+    id: 287,
+    name: 'Slakoth',
+    generation: PokemonGeneration.Gen3,
+    cry: 'Slakoth!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  vigoroth: {
+    id: 288,
+    name: 'Vigoroth',
+    generation: PokemonGeneration.Gen3,
+    cry: 'Vigoroth!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  slaking: {
+    id: 289,
+    name: 'Slaking',
+    generation: PokemonGeneration.Gen3,
+    cry: 'Slaking!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  nincada: {
+    id: 290,
+    name: 'Nincada',
+    generation: PokemonGeneration.Gen3,
+    cry: 'Nincada!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  ninjask: {
+    id: 291,
+    name: 'Ninjask',
+    generation: PokemonGeneration.Gen3,
+    cry: 'Ninjask!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  shedinja: {
+    id: 292,
+    name: 'Shedinja',
+    generation: PokemonGeneration.Gen3,
+    cry: 'Shedinja!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  whismur: {
+    id: 293,
+    name: 'Whismur',
+    generation: PokemonGeneration.Gen3,
+    cry: 'Whismur!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  loudred: {
+    id: 294,
+    name: 'Loudred',
+    generation: PokemonGeneration.Gen3,
+    cry: 'Loudred!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  exploud: {
+    id: 295,
+    name: 'Exploud',
+    generation: PokemonGeneration.Gen3,
+    cry: 'Exploud!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  makuhita: {
+    id: 296,
+    name: 'Makuhita',
+    generation: PokemonGeneration.Gen3,
+    cry: 'Makuhita!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  hariyama: {
+    id: 297,
+    name: 'Hariyama',
+    generation: PokemonGeneration.Gen3,
+    cry: 'Hariyama!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  azurill: {
+    id: 298,
+    name: 'Azurill',
+    generation: PokemonGeneration.Gen3,
+    cry: 'Azurill!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  nosepass: {
+    id: 299,
+    name: 'Nosepass',
+    generation: PokemonGeneration.Gen3,
+    cry: 'Nosepass!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  skitty: {
+    id: 300,
+    name: 'Skitty',
+    generation: PokemonGeneration.Gen3,
+    cry: 'Skitty!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  delcatty: {
+    id: 301,
+    name: 'Delcatty',
+    generation: PokemonGeneration.Gen3,
+    cry: 'Delcatty!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  sableye: {
+    id: 302,
+    name: 'Sableye',
+    generation: PokemonGeneration.Gen3,
+    cry: 'Sableye!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  mawile: {
+    id: 303,
+    name: 'Mawile',
+    generation: PokemonGeneration.Gen3,
+    cry: 'Mawile!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  aron: {
+    id: 304,
+    name: 'Aron',
+    generation: PokemonGeneration.Gen3,
+    cry: 'Aron!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  lairon: {
+    id: 305,
+    name: 'Lairon',
+    generation: PokemonGeneration.Gen3,
+    cry: 'Lairon!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  aggron: {
+    id: 306,
+    name: 'Aggron',
+    generation: PokemonGeneration.Gen3,
+    cry: 'Aggron!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  meditite: {
+    id: 307,
+    name: 'Meditite',
+    generation: PokemonGeneration.Gen3,
+    cry: 'Meditite!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  medicham: {
+    id: 308,
+    name: 'Medicham',
+    generation: PokemonGeneration.Gen3,
+    cry: 'Medicham!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  electrike: {
+    id: 309,
+    name: 'Electrike',
+    generation: PokemonGeneration.Gen3,
+    cry: 'Electrike!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  manectric: {
+    id: 310,
+    name: 'Manectric',
+    generation: PokemonGeneration.Gen3,
+    cry: 'Manectric!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  plusle: {
+    id: 311,
+    name: 'Plusle',
+    generation: PokemonGeneration.Gen3,
+    cry: 'Plusle!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  minun: {
+    id: 312,
+    name: 'Minun',
+    generation: PokemonGeneration.Gen3,
+    cry: 'Minun!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  volbeat: {
+    id: 313,
+    name: 'Volbeat',
+    generation: PokemonGeneration.Gen3,
+    cry: 'Volbeat!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  illumise: {
+    id: 314,
+    name: 'Illumise',
+    generation: PokemonGeneration.Gen3,
+    cry: 'Illumise!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  roselia: {
+    id: 315,
+    name: 'Roselia',
+    generation: PokemonGeneration.Gen3,
+    cry: 'Roselia!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  gulpin: {
+    id: 316,
+    name: 'Gulpin',
+    generation: PokemonGeneration.Gen3,
+    cry: 'Gulpin!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  swalot: {
+    id: 317,
+    name: 'Swalot',
+    generation: PokemonGeneration.Gen3,
+    cry: 'Swalot!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  carvanha: {
+    id: 318,
+    name: 'Carvanha',
+    generation: PokemonGeneration.Gen3,
+    cry: 'Carvanha!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  sharpedo: {
+    id: 319,
+    name: 'Sharpedo',
+    generation: PokemonGeneration.Gen3,
+    cry: 'Sharpedo!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  wailmer: {
+    id: 320,
+    name: 'Wailmer',
+    generation: PokemonGeneration.Gen3,
+    cry: 'Wailmer!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  wailord: {
+    id: 321,
+    name: 'Wailord',
+    generation: PokemonGeneration.Gen3,
+    cry: 'Wailord!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+    originalSpriteSize: 64,
+  },
+  numel: {
+    id: 322,
+    name: 'Numel',
+    generation: PokemonGeneration.Gen3,
+    cry: 'Numel!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  camerupt: {
+    id: 323,
+    name: 'Camerupt',
+    generation: PokemonGeneration.Gen3,
+    cry: 'Camerupt!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  torkoal: {
+    id: 324,
+    name: 'Torkoal',
+    generation: PokemonGeneration.Gen3,
+    cry: 'Torkoal!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  spoink: {
+    id: 325,
+    name: 'Spoink',
+    generation: PokemonGeneration.Gen3,
+    cry: 'Spoink!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  grumpig: {
+    id: 326,
+    name: 'Grumpig',
+    generation: PokemonGeneration.Gen3,
+    cry: 'Grumpig!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  spinda: {
+    id: 327,
+    name: 'Spinda',
+    generation: PokemonGeneration.Gen3,
+    cry: 'Spinda!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  trapinch: {
+    id: 328,
+    name: 'Trapinch',
+    generation: PokemonGeneration.Gen3,
+    cry: 'Trapinch!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  vibrava: {
+    id: 329,
+    name: 'Vibrava',
+    generation: PokemonGeneration.Gen3,
+    cry: 'Vibrava!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  flygon: {
+    id: 330,
+    name: 'Flygon',
+    generation: PokemonGeneration.Gen3,
+    cry: 'Flygon!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  cacnea: {
+    id: 331,
+    name: 'Cacnea',
+    generation: PokemonGeneration.Gen3,
+    cry: 'Cacnea!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  cacturne: {
+    id: 332,
+    name: 'Cacturne',
+    generation: PokemonGeneration.Gen3,
+    cry: 'Cacturne!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  swablu: {
+    id: 333,
+    name: 'Swablu',
+    generation: PokemonGeneration.Gen3,
+    cry: 'Swablu!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  altaria: {
+    id: 334,
+    name: 'Altaria',
+    generation: PokemonGeneration.Gen3,
+    cry: 'Altaria!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  zangoose: {
+    id: 335,
+    name: 'Zangoose',
+    generation: PokemonGeneration.Gen3,
+    cry: 'Zangoose!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  seviper: {
+    id: 336,
+    name: 'Seviper',
+    generation: PokemonGeneration.Gen3,
+    cry: 'Seviper!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  lunatone: {
+    id: 337,
+    name: 'Lunatone',
+    generation: PokemonGeneration.Gen3,
+    cry: 'Lunatone!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  solrock: {
+    id: 338,
+    name: 'Solrock',
+    generation: PokemonGeneration.Gen3,
+    cry: 'Solrock!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  barboach: {
+    id: 339,
+    name: 'Barboach',
+    generation: PokemonGeneration.Gen3,
+    cry: 'Barboach!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  whiscash: {
+    id: 340,
+    name: 'Whiscash',
+    generation: PokemonGeneration.Gen3,
+    cry: 'Whiscash!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  corphish: {
+    id: 341,
+    name: 'Corphish',
+    generation: PokemonGeneration.Gen3,
+    cry: 'Corphish!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  crawdaunt: {
+    id: 342,
+    name: 'Crawdaunt',
+    generation: PokemonGeneration.Gen3,
+    cry: 'Crawdaunt!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  baltoy: {
+    id: 343,
+    name: 'Baltoy',
+    generation: PokemonGeneration.Gen3,
+    cry: 'Baltoy!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  claydol: {
+    id: 344,
+    name: 'Claydol',
+    generation: PokemonGeneration.Gen3,
+    cry: 'Claydol!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  lileep: {
+    id: 345,
+    name: 'Lileep',
+    generation: PokemonGeneration.Gen3,
+    cry: 'Lileep!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  cradily: {
+    id: 346,
+    name: 'Cradily',
+    generation: PokemonGeneration.Gen3,
+    cry: 'Cradily!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  anorith: {
+    id: 347,
+    name: 'Anorith',
+    generation: PokemonGeneration.Gen3,
+    cry: 'Anorith!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  armaldo: {
+    id: 348,
+    name: 'Armaldo',
+    generation: PokemonGeneration.Gen3,
+    cry: 'Armaldo!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  feebas: {
+    id: 349,
+    name: 'Feebas',
+    generation: PokemonGeneration.Gen3,
+    cry: 'Feebas!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  milotic: {
+    id: 350,
+    name: 'Milotic',
+    generation: PokemonGeneration.Gen3,
+    cry: 'Milotic!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  castform: {
+    id: 351,
+    name: 'Castform',
+    generation: PokemonGeneration.Gen3,
+    cry: 'Castform!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  kecleon: {
+    id: 352,
+    name: 'Kecleon',
+    generation: PokemonGeneration.Gen3,
+    cry: 'Kecleon!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  shuppet: {
+    id: 353,
+    name: 'Shuppet',
+    generation: PokemonGeneration.Gen3,
+    cry: 'Shuppet!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  banette: {
+    id: 354,
+    name: 'Banette',
+    generation: PokemonGeneration.Gen3,
+    cry: 'Banette!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  duskull: {
+    id: 355,
+    name: 'Duskull',
+    generation: PokemonGeneration.Gen3,
+    cry: 'Duskull!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  dusclops: {
+    id: 356,
+    name: 'Dusclops',
+    generation: PokemonGeneration.Gen3,
+    cry: 'Dusclops!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  tropius: {
+    id: 357,
+    name: 'Tropius',
+    generation: PokemonGeneration.Gen3,
+    cry: 'Tropius!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  chimecho: {
+    id: 358,
+    name: 'Chimecho',
+    generation: PokemonGeneration.Gen3,
+    cry: 'Chimecho!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  absol: {
+    id: 359,
+    name: 'Absol',
+    generation: PokemonGeneration.Gen3,
+    cry: 'Absol!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  wynaut: {
+    id: 360,
+    name: 'Wynaut',
+    generation: PokemonGeneration.Gen3,
+    cry: 'Wynaut!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  snorunt: {
+    id: 361,
+    name: 'Snorunt',
+    generation: PokemonGeneration.Gen3,
+    cry: 'Snorunt!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  glalie: {
+    id: 362,
+    name: 'Glalie',
+    generation: PokemonGeneration.Gen3,
+    cry: 'Glalie!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  spheal: {
+    id: 363,
+    name: 'Spheal',
+    generation: PokemonGeneration.Gen3,
+    cry: 'Spheal!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  sealeo: {
+    id: 364,
+    name: 'Sealeo',
+    generation: PokemonGeneration.Gen3,
+    cry: 'Sealeo!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  walrein: {
+    id: 365,
+    name: 'Walrein',
+    generation: PokemonGeneration.Gen3,
+    cry: 'Walrein!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  clamperl: {
+    id: 366,
+    name: 'Clamperl',
+    generation: PokemonGeneration.Gen3,
+    cry: 'Clamperl!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  huntail: {
+    id: 367,
+    name: 'Huntail',
+    generation: PokemonGeneration.Gen3,
+    cry: 'Huntail!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  gorebyss: {
+    id: 368,
+    name: 'Gorebyss',
+    generation: PokemonGeneration.Gen3,
+    cry: 'Gorebyss!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  relicanth: {
+    id: 369,
+    name: 'Relicanth',
+    generation: PokemonGeneration.Gen3,
+    cry: 'Relicanth!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  luvdisc: {
+    id: 370,
+    name: 'Luvdisc',
+    generation: PokemonGeneration.Gen3,
+    cry: 'Luvdisc!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  bagon: {
+    id: 371,
+    name: 'Bagon',
+    generation: PokemonGeneration.Gen3,
+    cry: 'Bagon!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  shelgon: {
+    id: 372,
+    name: 'Shelgon',
+    generation: PokemonGeneration.Gen3,
+    cry: 'Shelgon!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  salamence: {
+    id: 373,
+    name: 'Salamence',
+    generation: PokemonGeneration.Gen3,
+    cry: 'Salamence!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  beldum: {
+    id: 374,
+    name: 'Beldum',
+    generation: PokemonGeneration.Gen3,
+    cry: 'Beldum!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  metang: {
+    id: 375,
+    name: 'Metang',
+    generation: PokemonGeneration.Gen3,
+    cry: 'Metang!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  metagross: {
+    id: 376,
+    name: 'Metagross',
+    generation: PokemonGeneration.Gen3,
+    cry: 'Metagross!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  regirock: {
+    id: 377,
+    name: 'Regirock',
+    generation: PokemonGeneration.Gen3,
+    cry: 'Regirock!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  regice: {
+    id: 378,
+    name: 'Regice',
+    generation: PokemonGeneration.Gen3,
+    cry: 'Regice!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  registeel: {
+    id: 379,
+    name: 'Registeel',
+    generation: PokemonGeneration.Gen3,
+    cry: 'Registeel!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  latias: {
+    id: 380,
+    name: 'Latias',
+    generation: PokemonGeneration.Gen3,
+    cry: 'Latias!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  latios: {
+    id: 381,
+    name: 'Latios',
+    generation: PokemonGeneration.Gen3,
+    cry: 'Latios!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  kyogre: {
+    id: 382,
+    name: 'Kyogre',
+    generation: PokemonGeneration.Gen3,
+    cry: 'Kyogre!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+    originalSpriteSize: 64,
+  },
+  groudon: {
+    id: 383,
+    name: 'Groudon',
+    generation: PokemonGeneration.Gen3,
+    cry: 'Groudon!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+    originalSpriteSize: 64,
+  },
+  rayquaza: {
+    id: 384,
+    name: 'Rayquaza',
+    generation: PokemonGeneration.Gen3,
+    cry: 'Rayquaza!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+    originalSpriteSize: 64,
+  },
+  jirachi: {
+    id: 385,
+    name: 'Jirachi',
+    generation: PokemonGeneration.Gen3,
+    cry: 'Jirachi!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  deoxys: {
+    id: 386,
+    name: 'Deoxys',
+    generation: PokemonGeneration.Gen3,
+    cry: 'Deoxys!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+    extraSprites: [PokemonExtraSprite.leftFacing],
+  },
+  deoxys_speed: {
+    id: 386,
+    name: 'Deoxys (Speed)',
+    generation: PokemonGeneration.Gen3,
+    cry: 'Deoxys!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+    extraSprites: [PokemonExtraSprite.leftFacing],
+  },
+  deoxys_attack: {
+    id: 386,
+    name: 'Deoxys (Attack)',
+    generation: PokemonGeneration.Gen3,
+    cry: 'Deoxys!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+    extraSprites: [PokemonExtraSprite.leftFacing],
+  },
+  deoxys_defense: {
+    id: 386,
+    name: 'Deoxys (Defense)',
+    generation: PokemonGeneration.Gen3,
+    cry: 'Deoxys!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+};
