+++ src/common/dex/gen4.ts
@@ -1,0 +1,1038 @@
+import {
+  PokemonConfig,
+  PokemonGeneration,
+  PokemonColor,
+  PokemonExtraSprite,
+} from '../types';
+
+export const GEN_4_DEX: { [key: string]: PokemonConfig } = {
+  turtwig: {
+    id: 387,
+    name: 'Turtwig',
+    generation: PokemonGeneration.Gen4,
+    cry: 'Turtwig!',
+    possibleColors: [PokemonColor.default],
+  },
+  grotle: {
+    id: 388,
+    name: 'Grotle',
+    generation: PokemonGeneration.Gen4,
+    cry: 'Grotle!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  torterra: {
+    id: 389,
+    name: 'Torterra',
+    generation: PokemonGeneration.Gen4,
+    cry: 'Torterra!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+    extraSprites: [PokemonExtraSprite.leftFacing],
+  },
+  chimchar: {
+    id: 390,
+    name: 'Chimchar',
+    generation: PokemonGeneration.Gen4,
+    cry: 'Chimchar!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  monferno: {
+    id: 391,
+    name: 'Monferno',
+    generation: PokemonGeneration.Gen4,
+    cry: 'Monferno!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  infernape: {
+    id: 392,
+    name: 'Infernape',
+    generation: PokemonGeneration.Gen4,
+    cry: 'Infernape!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  piplup: {
+    id: 393,
+    name: 'Piplup',
+    generation: PokemonGeneration.Gen4,
+    cry: 'Piplup!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  prinplup: {
+    id: 394,
+    name: 'Prinplup',
+    generation: PokemonGeneration.Gen4,
+    cry: 'Prinplup!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  empoleon: {
+    id: 395,
+    name: 'Empoleon',
+    generation: PokemonGeneration.Gen4,
+    cry: 'Empoleon!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  starly: {
+    id: 396,
+    name: 'Starly',
+    generation: PokemonGeneration.Gen4,
+    cry: 'Starly!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  staravia: {
+    id: 397,
+    name: 'Staravia',
+    generation: PokemonGeneration.Gen4,
+    cry: 'Staravia!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  staraptor: {
+    id: 398,
+    name: 'Staraptor',
+    generation: PokemonGeneration.Gen4,
+    cry: 'Staraptor!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  bidoof: {
+    id: 399,
+    name: 'Bidoof',
+    generation: PokemonGeneration.Gen4,
+    cry: 'Bidoof!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  bibarel: {
+    id: 400,
+    name: 'Bibarel',
+    generation: PokemonGeneration.Gen4,
+    cry: 'Bibarel!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  kricketot: {
+    id: 401,
+    name: 'Kricketot',
+    generation: PokemonGeneration.Gen4,
+    cry: 'Kricketot!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+    extraSprites: [PokemonExtraSprite.leftFacing],
+  },
+  kricketune: {
+    id: 402,
+    name: 'Kricketune',
+    generation: PokemonGeneration.Gen4,
+    cry: 'Kricketune!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+    extraSprites: [PokemonExtraSprite.leftFacing],
+  },
+  shinx: {
+    id: 403,
+    name: 'Shinx',
+    generation: PokemonGeneration.Gen4,
+    cry: 'Shinx!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  luxio: {
+    id: 404,
+    name: 'Luxio',
+    generation: PokemonGeneration.Gen4,
+    cry: 'Luxio!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  luxray: {
+    id: 405,
+    name: 'Luxray',
+    generation: PokemonGeneration.Gen4,
+    cry: 'Luxray!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  budew: {
+    id: 406,
+    name: 'Budew',
+    generation: PokemonGeneration.Gen4,
+    cry: 'Budew!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  roserade: {
+    id: 407,
+    name: 'Roserade',
+    generation: PokemonGeneration.Gen4,
+    cry: 'Roserade!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+    extraSprites: [PokemonExtraSprite.leftFacing],
+  },
+  cranidos: {
+    id: 408,
+    name: 'Cranidos',
+    generation: PokemonGeneration.Gen4,
+    cry: 'Cranidos!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  rampardos: {
+    id: 409,
+    name: 'Rampardos',
+    generation: PokemonGeneration.Gen4,
+    cry: 'Rampardos!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  shieldon: {
+    id: 410,
+    name: 'Shieldon',
+    generation: PokemonGeneration.Gen4,
+    cry: 'Shieldon!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  bastiodon: {
+    id: 411,
+    name: 'Bastiodon',
+    generation: PokemonGeneration.Gen4,
+    cry: 'Bastiodon!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  burmy_plant: {
+    id: 412,
+    name: 'Burmy (Plant)',
+    generation: PokemonGeneration.Gen4,
+    cry: 'Burmy!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  burmy_sandy: {
+    id: 412,
+    name: 'Burmy (Sandy)',
+    generation: PokemonGeneration.Gen4,
+    cry: 'Burmy!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  burmy_trash: {
+    id: 412,
+    name: 'Burmy (Trash)',
+    generation: PokemonGeneration.Gen4,
+    cry: 'Burmy!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  wormadam_plant: {
+    id: 413,
+    name: 'Wormadam (Plant)',
+    generation: PokemonGeneration.Gen4,
+    cry: 'Wormadam!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  wormadam_sandy: {
+    id: 413,
+    name: 'Wormadam (Sandy)',
+    generation: PokemonGeneration.Gen4,
+    cry: 'Wormadam!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  wormadam_trash: {
+    id: 413,
+    name: 'Wormadam (Trash)',
+    generation: PokemonGeneration.Gen4,
+    cry: 'Wormadam!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  mothim: {
+    id: 414,
+    name: 'Mothim',
+    generation: PokemonGeneration.Gen4,
+    cry: 'Mothim!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  combee_male: {
+    id: 415,
+    name: 'Combee (Male)',
+    generation: PokemonGeneration.Gen4,
+    cry: 'Combee!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  combee_female: {
+    id: 415,
+    name: 'Combee (Female)',
+    generation: PokemonGeneration.Gen4,
+    cry: 'Combee!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  vespiquen: {
+    id: 416,
+    name: 'Vespiquen',
+    generation: PokemonGeneration.Gen4,
+    cry: 'Vespiquen!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  pachirisu: {
+    id: 417,
+    name: 'Pachirisu',
+    generation: PokemonGeneration.Gen4,
+    cry: 'Pachirisu!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  buizel: {
+    id: 418,
+    name: 'Buizel',
+    generation: PokemonGeneration.Gen4,
+    cry: 'Buizel!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  floatzel: {
+    id: 419,
+    name: 'Floatzel',
+    generation: PokemonGeneration.Gen4,
+    cry: 'Floatzel!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  cherubi: {
+    id: 420,
+    name: 'Cherubi',
+    generation: PokemonGeneration.Gen4,
+    cry: 'Cherubi!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  cherrim: {
+    id: 421,
+    name: 'Cherrim',
+    generation: PokemonGeneration.Gen4,
+    cry: 'Cherrim!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  shellos_west: {
+    id: 422,
+    name: 'Shellos (West)',
+    generation: PokemonGeneration.Gen4,
+    cry: 'Shellos!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  shellos_east: {
+    id: 422,
+    name: 'Shellos (East)',
+    generation: PokemonGeneration.Gen4,
+    cry: 'Shellos!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  gastrodon_west: {
+    id: 423,
+    name: 'Gastrodon (West)',
+    generation: PokemonGeneration.Gen4,
+    cry: 'Gastrodon!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  gastrodon_east: {
+    id: 423,
+    name: 'Gastrodon (East)',
+    generation: PokemonGeneration.Gen4,
+    cry: 'Gastrodon!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  ambipom: {
+    id: 424,
+    name: 'Ambipom',
+    generation: PokemonGeneration.Gen4,
+    cry: 'Ambipom!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+    extraSprites: [PokemonExtraSprite.leftFacing],
+  },
+  drifloon: {
+    id: 425,
+    name: 'Drifloon',
+    generation: PokemonGeneration.Gen4,
+    cry: 'Drifloon!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  drifblim: {
+    id: 426,
+    name: 'Drifblim',
+    generation: PokemonGeneration.Gen4,
+    cry: 'Drifblim!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  buneary: {
+    id: 427,
+    name: 'Buneary',
+    generation: PokemonGeneration.Gen4,
+    cry: 'Buneary!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+    extraSprites: [PokemonExtraSprite.leftFacing],
+  },
+  lopunny: {
+    id: 428,
+    name: 'Lopunny',
+    generation: PokemonGeneration.Gen4,
+    cry: 'Lopunny!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  mismagius: {
+    id: 429,
+    name: 'Mismagius',
+    generation: PokemonGeneration.Gen4,
+    cry: 'Mismagius!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  honchkrow: {
+    id: 430,
+    name: 'Honchkrow',
+    generation: PokemonGeneration.Gen4,
+    cry: 'Honchkrow!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  glameow: {
+    id: 431,
+    name: 'Glameow',
+    generation: PokemonGeneration.Gen4,
+    cry: 'Glameow!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  purugly: {
+    id: 432,
+    name: 'Purugly',
+    generation: PokemonGeneration.Gen4,
+    cry: 'Purugly!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  chingling: {
+    id: 433,
+    name: 'Chingling',
+    generation: PokemonGeneration.Gen4,
+    cry: 'Chingling!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  stunky: {
+    id: 434,
+    name: 'Stunky',
+    generation: PokemonGeneration.Gen4,
+    cry: 'Stunky!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  skuntank: {
+    id: 435,
+    name: 'Skuntank',
+    generation: PokemonGeneration.Gen4,
+    cry: 'Skuntank!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  bronzor: {
+    id: 436,
+    name: 'Bronzor',
+    generation: PokemonGeneration.Gen4,
+    cry: 'Bronzor!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  bronzong: {
+    id: 437,
+    name: 'Bronzong',
+    generation: PokemonGeneration.Gen4,
+    cry: 'Bronzong!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  bonsly: {
+    id: 438,
+    name: 'Bonsly',
+    generation: PokemonGeneration.Gen4,
+    cry: 'Bonsly!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  mimeJr: {
+    id: 439,
+    name: 'Mime Jr.',
+    generation: PokemonGeneration.Gen4,
+    cry: 'Mime Jr.!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  happiny: {
+    id: 440,
+    name: 'Happiny',
+    generation: PokemonGeneration.Gen4,
+    cry: 'Happiny!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  chatot: {
+    id: 441,
+    name: 'Chatot',
+    generation: PokemonGeneration.Gen4,
+    cry: 'Chatot!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  spiritomb: {
+    id: 442,
+    name: 'Spiritomb',
+    generation: PokemonGeneration.Gen4,
+    cry: 'Spiritomb!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  gible: {
+    id: 443,
+    name: 'Gible',
+    generation: PokemonGeneration.Gen4,
+    cry: 'Gible!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+    extraSprites: [PokemonExtraSprite.leftFacing],
+  },
+  gible_female: {
+    id: 443,
+    name: 'Gible (Female)',
+    generation: PokemonGeneration.Gen4,
+    cry: 'Gible!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+    extraSprites: [PokemonExtraSprite.leftFacing],
+  },
+  gabite: {
+    id: 444,
+    name: 'Gabite',
+    generation: PokemonGeneration.Gen4,
+    cry: 'Gabite!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  gabite_female: {
+    id: 444,
+    name: 'Gabite (Female)',
+    generation: PokemonGeneration.Gen4,
+    cry: 'Gabite!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  garchomp: {
+    id: 445,
+    name: 'Garchomp',
+    generation: PokemonGeneration.Gen4,
+    cry: 'Garchomp!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+    extraSprites: [PokemonExtraSprite.leftFacing],
+  },
+  garchomp_female: {
+    id: 445,
+    name: 'Garchomp (Female)',
+    generation: PokemonGeneration.Gen4,
+    cry: 'Garchomp!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+    extraSprites: [PokemonExtraSprite.leftFacing],
+  },
+  munchlax: {
+    id: 446,
+    name: 'Munchlax',
+    generation: PokemonGeneration.Gen4,
+    cry: 'Munchlax!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  riolu: {
+    id: 447,
+    name: 'Riolu',
+    generation: PokemonGeneration.Gen4,
+    cry: 'Riolu!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+    extraSprites: [PokemonExtraSprite.leftFacing],
+  },
+  lucario: {
+    id: 448,
+    name: 'Lucario',
+    generation: PokemonGeneration.Gen4,
+    cry: 'Lucario!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  hippopotas_male: {
+    id: 449,
+    name: 'Hippopotas (Male)',
+    generation: PokemonGeneration.Gen4,
+    cry: 'Hippopotas!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  hippopotas_female: {
+    id: 449,
+    name: 'Hippopotas (Female)',
+    generation: PokemonGeneration.Gen4,
+    cry: 'Hippopotas!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  hippowdon_male: {
+    id: 450,
+    name: 'Hippowdon (Male)',
+    generation: PokemonGeneration.Gen4,
+    cry: 'Hippowdon!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  hippowdon_female: {
+    id: 450,
+    name: 'Hippowdon (Female)',
+    generation: PokemonGeneration.Gen4,
+    cry: 'Hippowdon!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  skorupi: {
+    id: 451,
+    name: 'Skorupi',
+    generation: PokemonGeneration.Gen4,
+    cry: 'Skorupi!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  drapion: {
+    id: 452,
+    name: 'Drapion',
+    generation: PokemonGeneration.Gen4,
+    cry: 'Drapion!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  croagunk: {
+    id: 453,
+    name: 'Croagunk',
+    generation: PokemonGeneration.Gen4,
+    cry: 'Croagunk!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  toxicroak: {
+    id: 454,
+    name: 'Toxicroak',
+    generation: PokemonGeneration.Gen4,
+    cry: 'Toxicroak!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  carnivine: {
+    id: 455,
+    name: 'Carnivine',
+    generation: PokemonGeneration.Gen4,
+    cry: 'Carnivine!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  finneon: {
+    id: 456,
+    name: 'Finneon',
+    generation: PokemonGeneration.Gen4,
+    cry: 'Finneon!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  lumineon: {
+    id: 457,
+    name: 'Lumineon',
+    generation: PokemonGeneration.Gen4,
+    cry: 'Lumineon!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  mantyke: {
+    id: 458,
+    name: 'Mantyke',
+    generation: PokemonGeneration.Gen4,
+    cry: 'Mantyke!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  snover: {
+    id: 459,
+    name: 'Snover',
+    generation: PokemonGeneration.Gen4,
+    cry: 'Snover!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  abomasnow: {
+    id: 460,
+    name: 'Abomasnow',
+    generation: PokemonGeneration.Gen4,
+    cry: 'Abomasnow!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  weavile: {
+    id: 461,
+    name: 'Weavile',
+    generation: PokemonGeneration.Gen4,
+    cry: 'Weavile!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  magnezone: {
+    id: 462,
+    name: 'Magnezone',
+    generation: PokemonGeneration.Gen4,
+    cry: 'Magnezone!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  lickilicky: {
+    id: 463,
+    name: 'Lickilicky',
+    generation: PokemonGeneration.Gen4,
+    cry: 'Lickilicky!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  rhyperior: {
+    id: 464,
+    name: 'Rhyperior',
+    generation: PokemonGeneration.Gen4,
+    cry: 'Rhyperior!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  tangrowth: {
+    id: 465,
+    name: 'Tangrowth',
+    generation: PokemonGeneration.Gen4,
+    cry: 'Tangrowth!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  electivire: {
+    id: 466,
+    name: 'Electivire',
+    generation: PokemonGeneration.Gen4,
+    cry: 'Electivire!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  magmortar: {
+    id: 467,
+    name: 'Magmortar',
+    generation: PokemonGeneration.Gen4,
+    cry: 'Magmortar!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  togekiss: {
+    id: 468,
+    name: 'Togekiss',
+    generation: PokemonGeneration.Gen4,
+    cry: 'Togekiss!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+    extraSprites: [PokemonExtraSprite.leftFacing],
+  },
+  yanmega: {
+    id: 469,
+    name: 'Yanmega',
+    generation: PokemonGeneration.Gen4,
+    cry: 'Yanmega!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  leafeon: {
+    id: 470,
+    name: 'Leafeon',
+    generation: PokemonGeneration.Gen4,
+    cry: 'Leafeon!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  glaceon: {
+    id: 471,
+    name: 'Glaceon',
+    generation: PokemonGeneration.Gen4,
+    cry: 'Glaceon!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  gliscor: {
+    id: 472,
+    name: 'Gliscor',
+    generation: PokemonGeneration.Gen4,
+    cry: 'Gliscor!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  mamoswine: {
+    id: 473,
+    name: 'Mamoswine',
+    generation: PokemonGeneration.Gen4,
+    cry: 'Mamoswine!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  porygonZ: {
+    id: 474,
+    name: 'Porygon-Z',
+    generation: PokemonGeneration.Gen4,
+    cry: 'Porygon-Z!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  gallade: {
+    id: 475,
+    name: 'Gallade',
+    generation: PokemonGeneration.Gen4,
+    cry: 'Gallade!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  probopass: {
+    id: 476,
+    name: 'Probopass',
+    generation: PokemonGeneration.Gen4,
+    cry: 'Probopass!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  dusknoir: {
+    id: 477,
+    name: 'Dusknoir',
+    generation: PokemonGeneration.Gen4,
+    cry: 'Dusknoir!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  froslass: {
+    id: 478,
+    name: 'Froslass',
+    generation: PokemonGeneration.Gen4,
+    cry: 'Froslass!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  rotom: {
+    id: 479,
+    name: 'Rotom',
+    generation: PokemonGeneration.Gen4,
+    cry: 'Rotom!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  rotom_heat: {
+    id: 479,
+    name: 'Rotom (Heat Form)',
+    generation: PokemonGeneration.Gen4,
+    cry: 'Rotom!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  rotom_wash: {
+    id: 479,
+    name: 'Rotom (Wash Form)',
+    generation: PokemonGeneration.Gen4,
+    cry: 'Rotom!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+    extraSprites: [PokemonExtraSprite.leftFacing],
+  },
+  rotom_frost: {
+    id: 479,
+    name: 'Rotom (Frost Form)',
+    generation: PokemonGeneration.Gen4,
+    cry: 'Rotom!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  rotom_fan: {
+    id: 479,
+    name: 'Rotom (Fan Form)',
+    generation: PokemonGeneration.Gen4,
+    cry: 'Rotom!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  rotom_mow: {
+    id: 479,
+    name: 'Rotom (Mow Form)',
+    generation: PokemonGeneration.Gen4,
+    cry: 'Rotom!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  uxie: {
+    id: 480,
+    name: 'Uxie',
+    generation: PokemonGeneration.Gen4,
+    cry: 'Uxie!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  mesprit: {
+    id: 481,
+    name: 'Mesprit',
+    generation: PokemonGeneration.Gen4,
+    cry: 'Mesprit!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  azelf: {
+    id: 482,
+    name: 'Azelf',
+    generation: PokemonGeneration.Gen4,
+    cry: 'Azelf!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  dialga: {
+    id: 483,
+    name: 'Dialga',
+    generation: PokemonGeneration.Gen4,
+    cry: 'Dialga!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+    originalSpriteSize: 64,
+  },
+  palkia: {
+    id: 484,
+    name: 'Palkia',
+    generation: PokemonGeneration.Gen4,
+    cry: 'Palkia!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+    originalSpriteSize: 64,
+  },
+  heatran: {
+    id: 485,
+    name: 'Heatran',
+    generation: PokemonGeneration.Gen4,
+    cry: 'Heatran!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  regigigas: {
+    id: 486,
+    name: 'Regigigas',
+    generation: PokemonGeneration.Gen4,
+    cry: 'Regigigas!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+    originalSpriteSize: 64,
+  },
+  giratina: {
+    id: 487,
+    name: 'Giratina',
+    generation: PokemonGeneration.Gen4,
+    cry: 'Giratina!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+    originalSpriteSize: 64,
+  },
+  giratina_origin: {
+    id: 487,
+    name: 'Giratina (Origin Form)',
+    generation: PokemonGeneration.Gen4,
+    cry: 'Giratina!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+    originalSpriteSize: 64,
+  },
+  cresselia: {
+    id: 488,
+    name: 'Cresselia',
+    generation: PokemonGeneration.Gen4,
+    cry: 'Cresselia!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  phione: {
+    id: 489,
+    name: 'Phione',
+    generation: PokemonGeneration.Gen4,
+    cry: 'Phione!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  manaphy: {
+    id: 490,
+    name: 'Manaphy',
+    generation: PokemonGeneration.Gen4,
+    cry: 'Manaphy!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  darkrai: {
+    id: 491,
+    name: 'Darkrai',
+    generation: PokemonGeneration.Gen4,
+    cry: 'Darkrai!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  shaymin: {
+    id: 492,
+    name: 'Shaymin',
+    generation: PokemonGeneration.Gen4,
+    cry: 'Shaymin!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  shaymin_sky: {
+    id: 492,
+    name: 'Shaymin (Sky Form)',
+    generation: PokemonGeneration.Gen4,
+    cry: 'Shaymin!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+    extraSprites: [PokemonExtraSprite.leftFacing],
+  },
+  arceus: {
+    id: 493,
+    name: 'Arceus',
+    generation: PokemonGeneration.Gen4,
+    cry: 'Arceus!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+    originalSpriteSize: 64,
+  },
+  arceus_bug: {
+    id: 493,
+    name: 'Arceus (Bug)',
+    generation: PokemonGeneration.Gen4,
+    cry: 'Arceus!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+  },
+  arceus_dark: {
+    id: 493,
+    name: 'Arceus (Dark)',
+    generation: PokemonGeneration.Gen4,
+    cry: 'Arceus!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+    originalSpriteSize: 64,
+  },
+  arceus_dragon: {
+    id: 493,
+    name: 'Arceus (Dragon)',
+    generation: PokemonGeneration.Gen4,
+    cry: 'Arceus!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+    originalSpriteSize: 64,
+  },
+  arceus_electric: {
+    id: 493,
+    name: 'Arceus (Electric)',
+    generation: PokemonGeneration.Gen4,
+    cry: 'Arceus!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+    originalSpriteSize: 64,
+  },
+  arceus_fighting: {
+    id: 493,
+    name: 'Arceus (Fighting)',
+    generation: PokemonGeneration.Gen4,
+    cry: 'Arceus!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+    originalSpriteSize: 64,
+  },
+  arceus_fire: {
+    id: 493,
+    name: 'Arceus (Fire)',
+    generation: PokemonGeneration.Gen4,
+    cry: 'Arceus!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+    originalSpriteSize: 64,
+  },
+  arceus_flying: {
+    id: 493,
+    name: 'Arceus (Flying)',
+    generation: PokemonGeneration.Gen4,
+    cry: 'Arceus!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+    originalSpriteSize: 64,
+  },
+  arceus_ghost: {
+    id: 493,
+    name: 'Arceus (Ghost)',
+    generation: PokemonGeneration.Gen4,
+    cry: 'Arceus!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+    originalSpriteSize: 64,
+  },
+  arceus_grass: {
+    id: 493,
+    name: 'Arceus (Grass)',
+    generation: PokemonGeneration.Gen4,
+    cry: 'Arceus!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+    originalSpriteSize: 64,
+  },
+  arceus_ground: {
+    id: 493,
+    name: 'Arceus (Ground)',
+    generation: PokemonGeneration.Gen4,
+    cry: 'Arceus!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+    originalSpriteSize: 64,
+  },
+  arceus_ice: {
+    id: 493,
+    name: 'Arceus (Ice)',
+    generation: PokemonGeneration.Gen4,
+    cry: 'Arceus!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+    originalSpriteSize: 64,
+  },
+  arceus_poison: {
+    id: 493,
+    name: 'Arceus (Poison)',
+    generation: PokemonGeneration.Gen4,
+    cry: 'Arceus!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+    originalSpriteSize: 64,
+  },
+  arceus_psychic: {
+    id: 493,
+    name: 'Arceus (Psychic)',
+    generation: PokemonGeneration.Gen4,
+    cry: 'Arceus!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+    originalSpriteSize: 64,
+  },
+  arceus_rock: {
+    id: 493,
+    name: 'Arceus (Rock)',
+    generation: PokemonGeneration.Gen4,
+    cry: 'Arceus!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+    originalSpriteSize: 64,
+  },
+  arceus_steel: {
+    id: 493,
+    name: 'Arceus (Steel)',
+    generation: PokemonGeneration.Gen4,
+    cry: 'Arceus!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+    originalSpriteSize: 64,
+  },
+  arceus_water: {
+    id: 493,
+    name: 'Arceus (Water)',
+    generation: PokemonGeneration.Gen4,
+    cry: 'Arceus!',
+    possibleColors: [PokemonColor.default, PokemonColor.shiny],
+    originalSpriteSize: 64,
+  },
+};
